+++ lib/qsu.dart
@@ -1,6 +1,7 @@
 library;
 
 export 'src/array.dart';
+export 'src/crypto.dart';
 export 'src/format.dart';
 export 'src/misc.dart';
 export 'src/string.dart';