+++ lib/src/crypto.dart
@@ -5,3 +5,7 @@ import 'package:crypto/crypto.dart';
 String md5Hash(String str) {
   return md5.convert(utf8.encode(str)).toString();
 }
+
+String sha1Hash(String str) {
+  return sha1.convert(utf8.encode(str)).toString();
+}